Moved UserInvitationView and DomainRequestView
